requirements.txt
Django==4.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
